add nvtx getters with dataclass, restore string funcs
since we now just input the int ids to the dataclasses, i have added the
string functions as dataclass methods.
nvtx functions allow you to retrieve nvtx events from the table, and
then get the projection, which returns all the kernels in that zone.
then to get the nvtx duration it should be the same as last - first
(which can be done due to the new kernel __sub__)